Removing userId in Like | Validation user current in CommentController

+++ src/controllers/CommentController.ts
@@ -1,3 +1,4 @@
+import { ForbiddenError } from 'apollo-server';
 import PostRepository from '../repositories/PostRepository';
 import CommentModel from '../models/CommentModel';
 
@@ -17,7 +18,18 @@ class CommentController {
     return post!.comments;
   }
 
-  async updateComment(_: any, args: any) {
+  async updateComment(_: any, args: any, context: any) {
+    const comment = await this.postRepository.findComment(
+      args.input.postId,
+      args.input.commentId
+    );
+
+    if (context.user._id !== comment!.userId.toString()) {
+      throw new ForbiddenError(
+        'Authenticated user is not the author of the text'
+      );
+    }
+
     await this.postRepository.updateComment(
       args.input.postId,
       args.input.commentId,
@@ -28,7 +40,18 @@ class CommentController {
     return post!.comments;
   }
 
-  async deleteComment(_: any, args: any) {
+  async deleteComment(_: any, args: any, context: any) {
+    const comment = await this.postRepository.findComment(
+      args.input.postId,
+      args.input.commentId
+    );
+
+    if (context.user._id !== comment!.userId.toString()) {
+      throw new ForbiddenError(
+        'Authenticated user is not the author of the text'
+      );
+    }
+
     await this.postRepository.deleteComment(
       args.input.postId,
       args.input.commentId
